backend/requirements-pi.txt
smbus2
gpiozero
serial
rpi_ws281x
adafruit-circuitpython-neopixel
pvporcupine
sounddevice
vosk
requests
